+++ src/main/kotlin/camp/nextstep/blackjack/CardShuffler.kt
@@ -1,5 +1,7 @@
 package camp.nextstep.blackjack
 
+import camp.nextstep.blackjack.card.CardDeck
+
 object CardShuffler {
 
     fun shuffle(cardDeck: CardDeck): CardDeck {
