+++ src/main/kotlin/camp/nextstep/blackjack/Player.kt
@@ -1,5 +1,7 @@
 package camp.nextstep.blackjack
 
+import camp.nextstep.blackjack.card.Card
+
 class Player(val name: String) {
     private val _cards = mutableListOf<Card>()
     val cards get() = _cards.toList()
