+++ src/test/kotlin/camp/nextstep/blackjack/CardShufflerTest.kt
@@ -1,5 +1,6 @@
 package camp.nextstep.blackjack
 
+import camp.nextstep.blackjack.card.CardDeck
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
